16. CPython #1: PyBytesObject

+++ 0x04-python-more_data_structures/103-python.c
@@ -7,7 +7,7 @@
  */
 void print_python_bytes(PyObject *p)
 {
-	int length, idx, max;
+	long int length, idx, max;
 	char *str;
 
 	printf("[.] bytes object info\n");
@@ -18,13 +18,13 @@ void print_python_bytes(PyObject *p)
 	}
 	length = ((PyVarObject *)(p))->ob_size;
 	str = ((PyBytesObject *)p)->ob_sval;
-	printf("  size: %d\n", length);
+	printf("  size: %ld\n", length);
 	printf("  trying string: %s\n", str);
 	if (length >= 10)
 		max = 10;
 	else
 		max = length + 1;
-	printf("  first %d bytes:", max);
+	printf("  first %ld bytes:", max);
 	for (idx = 0; idx < max; idx++)
 		if (str[idx] >= 0)
 			printf(" %02x", str[idx]);
@@ -40,18 +40,18 @@ void print_python_bytes(PyObject *p)
 void print_python_list(PyObject *p)
 {
 	PyListObject *temp;
-	int length, idx;
+	long int length, idx;
 	PyObject *ptr;
 
 	length = ((PyVarObject *)(p))->ob_size;
 	temp = (PyListObject *)p;
 	printf("[*] Python list info\n");
-	printf("[*] Size of the Python List = %d\n", length);
-	printf("[*] Allocated = %d\n", temp->allocated);
+	printf("[*] Size of the Python List = %ld\n", length);
+	printf("[*] Allocated = %ld\n", temp->allocated);
 	for (idx = 0; idx < length; idx++)
 	{
 		ptr = ((PyListObject *)p)->ob_item[idx];
-		printf("Element %d: %s\n", idx, ((ptr)->ob_type)->tp_name);
+		printf("Element %ld: %s\n", idx, ((ptr)->ob_type)->tp_name);
 		if (PyBytes_Check(ptr))
 			print_python_bytes(ptr);
 	}
